% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm-classify.R
\name{llm_classify}
\alias{llm_classify}
\alias{llm_vec_classify}
\title{Categorize data as one of options given}
\usage{
llm_classify(
  .data,
  col,
  labels,
  pred_name = ".classify",
  additional_prompt = ""
)

llm_vec_classify(x, labels, additional_prompt = "")
}
\arguments{
\item{.data}{A \code{data.frame} or \code{tbl} object that contains the text to be
analyzed}

\item{col}{The name of the field to analyze, supports \code{tidy-eval}}

\item{labels}{A character vector with at least 2 labels to classify the text
as}

\item{pred_name}{A character vector with the name of the new column where the
prediction will be placed}

\item{additional_prompt}{Inserts this text into the prompt sent to the LLM}

\item{x}{A vector that contains the text to be analyzed}
}
\value{
\code{llm_classify} returns a \code{data.frame} or \code{tbl} object.
\code{llm_vec_classify} returns a vector that is the same length as \code{x}.
}
\description{
Use a Large Language Model (LLM) to classify the provided text as one of the
options provided via the \code{labels} argument.
}
