% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m-backend-generate.R, R/m-backend-prompt.R
\name{m_backend_generate}
\alias{m_backend_generate}
\alias{m_backend_prompt}
\title{Functions to integrate different back-ends}
\usage{
m_backend_generate(backend, x, base_prompt)

m_backend_prompt(backend, additional)
}
\arguments{
\item{backend}{An \code{mall_defaults} object}

\item{x}{The body of the text to be submitted to the LLM}

\item{base_prompt}{The instructions to the LLM about what to do with \code{x}}

\item{additional}{Additional text to insert to the \code{base_prompt}}
}
\value{
\code{m_backend_generate} does not return an object. \code{m_backend_prompt}
returns a list of functions that contain the base prompts.
}
\description{
Functions to integrate different back-ends
}
\keyword{internal}
