% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m-backend-prompt.R, R/m-backend-submit.R
\name{m_backend_prompt}
\alias{m_backend_prompt}
\alias{m_backend_submit}
\title{Functions to integrate different back-ends}
\usage{
m_backend_prompt(backend, additional)

m_backend_submit(backend, x, prompt, preview = FALSE)
}
\arguments{
\item{backend}{An \code{mall_session} object}

\item{additional}{Additional text to insert to the \code{base_prompt}}

\item{x}{The body of the text to be submitted to the LLM}

\item{prompt}{The additional information to add to the submission}

\item{preview}{If \code{TRUE}, it will display the resulting R call of the
first text in \code{x}}
}
\value{
\code{m_backend_submit} does not return an object. \code{m_backend_prompt}
returns a list of functions that contain the base prompts.
}
\description{
Functions to integrate different back-ends
}
\keyword{internal}
