% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-reviews.R
\docType{data}
\name{reviews}
\alias{reviews}
\title{Mini reviews data set}
\format{
A data frame that contains 3 records. The records are of fictitious
product reviews.
}
\usage{
reviews
}
\description{
Mini reviews data set
}
\examples{
library(mall)
data(reviews)
reviews

}
\keyword{datasets}
