% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm-use.R
\name{llm_use}
\alias{llm_use}
\title{Specify the model to use}
\usage{
llm_use(backend = NULL, model = NULL, ..., .silent = FALSE, force = FALSE)
}
\arguments{
\item{backend}{The name of an supported back-end provider. Currently only
'ollama' is supported.}

\item{model}{The name of model supported by the back-end provider}

\item{...}{Additional arguments that this function will pass down to the
integrating function. In the case of Ollama, it will pass those argument to
\code{ollamar::generate()}.}

\item{.silent}{Avoids console output}

\item{force}{Flag that tell the function to reset all of the settings in the
R session}
}
\value{
A \code{mall_defaults} object
}
\description{
Allows us to specify the back-end provider, model to use during the current
R session
}
